+++ noodles-util/src/variant.rs
@@ -4,3 +4,6 @@
 pub mod r#async;
 
 pub mod io;
+mod record;
+
+pub use self::record::Record;